+++ src/fn_graph.rs
@@ -14,11 +14,10 @@ use crate::{Edge, EdgeCounts, FnId, FnIdInner, Rank};
 #[cfg(feature = "async")]
 use daggy::Walker;
 #[cfg(feature = "async")]
-use futures::task::Poll;
-#[cfg(feature = "async")]
 use futures::{
-    future::{BoxFuture, LocalBoxFuture},
+    future::{Future, LocalBoxFuture},
     stream::{self, Stream, StreamExt},
+    task::Poll,
 };
 #[cfg(feature = "async")]
 use tokio::sync::{
@@ -200,7 +199,7 @@ impl<F> FnGraph<F> {
     /// Functions are produced by the stream only when all of their predecessors
     /// have returned.
     #[cfg(feature = "async")]
-    pub async fn fold_async<'f, Seed, FnFold>(&mut self, seed: Seed, fn_fold: FnFold) -> Seed
+    pub async fn fold_async<Seed, FnFold>(&mut self, seed: Seed, fn_fold: FnFold) -> Seed
     where
         FnFold: FnMut(Seed, &mut F) -> LocalBoxFuture<'_, Seed>,
     {
@@ -212,7 +211,7 @@ impl<F> FnGraph<F> {
     /// Functions are produced by the stream only when all of their predecessors
     /// have returned.
     #[cfg(feature = "async")]
-    pub async fn fold_rev_async<'f, Seed, FnFold>(&mut self, seed: Seed, fn_fold: FnFold) -> Seed
+    pub async fn fold_rev_async<Seed, FnFold>(&mut self, seed: Seed, fn_fold: FnFold) -> Seed
     where
         FnFold: FnMut(Seed, &mut F) -> LocalBoxFuture<'_, Seed>,
     {
@@ -220,7 +219,7 @@ impl<F> FnGraph<F> {
     }
 
     #[cfg(feature = "async")]
-    async fn fold_async_internal<'f, Seed, FnFold>(
+    async fn fold_async_internal<Seed, FnFold>(
         &mut self,
         seed: Seed,
         fn_fold: FnFold,
@@ -301,12 +300,14 @@ impl<F> FnGraph<F> {
     /// **Note:** a limit of zero is interpreted as no limit at all, and will
     /// have the same result as passing in `None`.
     #[cfg(feature = "async")]
-    pub async fn for_each_concurrent<FnForEach>(
-        &mut self,
+    pub async fn for_each_concurrent<'f, FnForEach, Fut>(
+        &'f self,
         limit: impl Into<Option<usize>>,
         fn_for_each: FnForEach,
     ) where
-        FnForEach: Fn(&mut F) -> BoxFuture<'_, ()>,
+        FnForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = ()> + 'f,
+        F: 'f,
     {
         self.for_each_concurrent_internal(limit, fn_for_each, IterDirection::Forward)
             .await
@@ -324,12 +325,14 @@ impl<F> FnGraph<F> {
     /// **Note:** a limit of zero is interpreted as no limit at all, and will
     /// have the same result as passing in `None`.
     #[cfg(feature = "async")]
-    pub async fn for_each_concurrent_rev<FnForEach>(
-        &mut self,
+    pub async fn for_each_concurrent_rev<'f, FnForEach, Fut>(
+        &'f self,
         limit: impl Into<Option<usize>>,
         fn_for_each: FnForEach,
     ) where
-        FnForEach: Fn(&mut F) -> BoxFuture<'_, ()>,
+        FnForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = ()> + 'f,
+        F: 'f,
     {
         self.for_each_concurrent_internal(limit, fn_for_each, IterDirection::Reverse)
             .await
@@ -337,13 +340,130 @@ impl<F> FnGraph<F> {
 
     // https://users.rust-lang.org/t/lifetime-may-not-live-long-enough-for-an-async-closure/62489
     #[cfg(feature = "async")]
-    async fn for_each_concurrent_internal<FnForEach>(
+    async fn for_each_concurrent_internal<'f, FnForEach, Fut>(
+        &'f self,
+        limit: impl Into<Option<usize>>,
+        fn_for_each: FnForEach,
+        iter_direction: IterDirection,
+    ) where
+        FnForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = ()> + 'f,
+        F: 'f,
+    {
+        let (graph_structure, predecessor_counts) = match iter_direction {
+            IterDirection::Forward => (&self.graph_structure, self.edge_counts.incoming().to_vec()),
+            IterDirection::Reverse => (
+                &self.graph_structure_rev,
+                self.edge_counts.outgoing().to_vec(),
+            ),
+        };
+        let channel_capacity = std::cmp::max(1, graph_structure.node_count());
+        let (fn_ready_tx, mut fn_ready_rx) = mpsc::channel(channel_capacity);
+        let (fn_done_tx, fn_done_rx) = mpsc::channel::<FnId>(channel_capacity);
+
+        // Preload the channel with all of the functions that have no predecessors
+        Topo::new(graph_structure)
+            .iter(graph_structure)
+            .filter(|fn_id| predecessor_counts[fn_id.index()] == 0)
+            .try_for_each(|fn_id| fn_ready_tx.try_send(fn_id))
+            .expect("Failed to preload function with no predecessors.");
+
+        let queuer =
+            Self::fn_ready_queuer(graph_structure, predecessor_counts, fn_done_rx, fn_ready_tx);
+
+        let fn_done_tx = RwLock::new(Some(fn_done_tx));
+        let fn_done_tx = &fn_done_tx;
+        let fn_for_each = &fn_for_each;
+        let fns_remaining = graph_structure.node_count();
+        let fns_remaining = RwLock::new(fns_remaining);
+        let fns_remaining = &fns_remaining;
+        let fn_refs = &self.graph;
+        let scheduler = async move {
+            stream::poll_fn(move |context| fn_ready_rx.poll_recv(context))
+                .for_each_concurrent(limit, |fn_id| async move {
+                    let r#fn = fn_refs.node_weight(fn_id).expect("Expected to borrow fn.");
+                    fn_for_each(r#fn).await;
+                    if let Some(fn_done_tx) = fn_done_tx.read().await.as_ref() {
+                        fn_done_tx
+                            .send(fn_id)
+                            .await
+                            .expect("Scheduler failed to send fn_id in `fn_done_tx`.");
+                    }
+
+                    // Close `fn_done_rx` when all functions have been executed,
+                    let fns_remaining_val = {
+                        let mut fns_remaining_ref = fns_remaining.write().await;
+                        *fns_remaining_ref -= 1;
+                        *fns_remaining_ref
+                    };
+                    if fns_remaining_val == 0 {
+                        fn_done_tx.write().await.take();
+                    }
+                })
+                .await;
+        };
+
+        futures::join!(queuer, scheduler);
+    }
+
+    /// Runs the provided logic over the functions concurrently in topological
+    /// order.
+    ///
+    /// The first argument is an optional `limit` on the number of concurrent
+    /// futures. If this limit is not `None`, no more than `limit` futures will
+    /// be run concurrently. The `limit` argument is of type
+    /// `Into<Option<usize>>`, and so can be provided as either `None`,
+    /// `Some(10)`, or just `10`.
+    ///
+    /// **Note:** a limit of zero is interpreted as no limit at all, and will
+    /// have the same result as passing in `None`.
+    #[cfg(feature = "async")]
+    pub async fn for_each_concurrent_mut<FnForEach, Fut>(
+        &mut self,
+        limit: impl Into<Option<usize>>,
+        fn_for_each: FnForEach,
+    ) where
+        FnForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = ()>,
+    {
+        self.for_each_concurrent_mut_internal(limit, fn_for_each, IterDirection::Forward)
+            .await
+    }
+
+    /// Runs the provided logic over the functions concurrently in reverse
+    /// topological order.
+    ///
+    /// The first argument is an optional `limit` on the number of concurrent
+    /// futures. If this limit is not `None`, no more than `limit` futures will
+    /// be run concurrently. The `limit` argument is of type
+    /// `Into<Option<usize>>`, and so can be provided as either `None`,
+    /// `Some(10)`, or just `10`.
+    ///
+    /// **Note:** a limit of zero is interpreted as no limit at all, and will
+    /// have the same result as passing in `None`.
+    #[cfg(feature = "async")]
+    pub async fn for_each_concurrent_mut_rev<FnForEach, Fut>(
+        &mut self,
+        limit: impl Into<Option<usize>>,
+        fn_for_each: FnForEach,
+    ) where
+        FnForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = ()>,
+    {
+        self.for_each_concurrent_mut_internal(limit, fn_for_each, IterDirection::Reverse)
+            .await
+    }
+
+    // https://users.rust-lang.org/t/lifetime-may-not-live-long-enough-for-an-async-closure/62489
+    #[cfg(feature = "async")]
+    async fn for_each_concurrent_mut_internal<FnForEach, Fut>(
         &mut self,
         limit: impl Into<Option<usize>>,
         fn_for_each: FnForEach,
         iter_direction: IterDirection,
     ) where
-        FnForEach: Fn(&mut F) -> BoxFuture<'_, ()>,
+        FnForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = ()>,
     {
         let (graph_structure, predecessor_counts) = match iter_direction {
             IterDirection::Forward => (&self.graph_structure, self.edge_counts.incoming().to_vec()),
@@ -468,14 +588,15 @@ impl<F> FnGraph<F> {
     /// **Note:** a limit of zero is interpreted as no limit at all, and will
     /// have the same result as passing in `None`.
     #[cfg(feature = "async")]
-    pub async fn try_for_each_concurrent<E, FnTryForEach>(
-        &mut self,
+    pub async fn try_for_each_concurrent<'f, E, FnTryForEach, Fut>(
+        &'f self,
         limit: impl Into<Option<usize>>,
         fn_try_for_each: FnTryForEach,
     ) -> Result<(), Vec<E>>
     where
         E: Debug,
-        FnTryForEach: Fn(&mut F) -> BoxFuture<'_, Result<(), E>>,
+        FnTryForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = Result<(), E>> + 'f,
     {
         self.try_for_each_concurrent_internal(limit, fn_try_for_each, IterDirection::Forward)
             .await
@@ -497,14 +618,15 @@ impl<F> FnGraph<F> {
     /// **Note:** a limit of zero is interpreted as no limit at all, and will
     /// have the same result as passing in `None`.
     #[cfg(feature = "async")]
-    pub async fn try_for_each_concurrent_rev<E, FnTryForEach>(
-        &mut self,
+    pub async fn try_for_each_concurrent_rev<'f, E, FnTryForEach, Fut>(
+        &'f self,
         limit: impl Into<Option<usize>>,
         fn_try_for_each: FnTryForEach,
     ) -> Result<(), Vec<E>>
     where
         E: Debug,
-        FnTryForEach: Fn(&mut F) -> BoxFuture<'_, Result<(), E>>,
+        FnTryForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = Result<(), E>> + 'f,
     {
         self.try_for_each_concurrent_internal(limit, fn_try_for_each, IterDirection::Reverse)
             .await
@@ -512,7 +634,161 @@ impl<F> FnGraph<F> {
 
     // https://users.rust-lang.org/t/lifetime-may-not-live-long-enough-for-an-async-closure/62489
     #[cfg(feature = "async")]
-    async fn try_for_each_concurrent_internal<E, FnTryForEach>(
+    async fn try_for_each_concurrent_internal<'f, E, FnTryForEach, Fut>(
+        &'f self,
+        limit: impl Into<Option<usize>>,
+        fn_try_for_each: FnTryForEach,
+        iter_direction: IterDirection,
+    ) -> Result<(), Vec<E>>
+    where
+        E: Debug,
+        FnTryForEach: Fn(&'f F) -> Fut,
+        Fut: Future<Output = Result<(), E>> + 'f,
+    {
+        let (graph_structure, predecessor_counts) = match iter_direction {
+            IterDirection::Forward => (&self.graph_structure, self.edge_counts.incoming().to_vec()),
+            IterDirection::Reverse => (
+                &self.graph_structure_rev,
+                self.edge_counts.outgoing().to_vec(),
+            ),
+        };
+        let channel_capacity = std::cmp::max(1, graph_structure.node_count());
+        let (result_tx, mut result_rx) = mpsc::channel(channel_capacity);
+        let (fn_ready_tx, mut fn_ready_rx) = mpsc::channel(channel_capacity);
+        let (fn_done_tx, fn_done_rx) = mpsc::channel::<FnId>(channel_capacity);
+
+        // Preload the channel with all of the functions that have no predecessors
+        Topo::new(graph_structure)
+            .iter(graph_structure)
+            .filter(|fn_id| predecessor_counts[fn_id.index()] == 0)
+            .try_for_each(|fn_id| fn_ready_tx.try_send(fn_id))
+            .expect("Failed to preload function with no predecessors.");
+
+        let queuer =
+            Self::fn_ready_queuer(graph_structure, predecessor_counts, fn_done_rx, fn_ready_tx);
+
+        let fn_done_tx = RwLock::new(Some(fn_done_tx));
+        let fn_done_tx = &fn_done_tx;
+        let fn_try_for_each = &fn_try_for_each;
+        let fns_remaining = graph_structure.node_count();
+        let fns_remaining = RwLock::new(fns_remaining);
+        let fns_remaining = &fns_remaining;
+        let fn_refs = &self.graph;
+        let scheduler = async move {
+            let result_tx_ref = &result_tx;
+
+            stream::poll_fn(move |context| fn_ready_rx.poll_recv(context))
+                .for_each_concurrent(limit, |fn_id| async move {
+                    let r#fn = fn_refs.node_weight(fn_id).expect("Expected to borrow fn.");
+                    if let Err(e) = fn_try_for_each(r#fn).await {
+                        result_tx_ref
+                            .send(e)
+                            .await
+                            .expect("Scheduler failed to send Err result in `result_tx`.");
+
+                        // Close `fn_done_rx`, which means `fn_ready_queuer` should return
+                        // `Poll::Ready(None)`.
+                        fn_done_tx.write().await.take();
+                    };
+
+                    if let Some(fn_done_tx) = fn_done_tx.read().await.as_ref() {
+                        fn_done_tx
+                            .send(fn_id)
+                            .await
+                            .expect("Scheduler failed to send fn_id in `fn_done_tx`.");
+                    }
+
+                    // Close `fn_done_rx` when all functions have been executed,
+                    let fns_remaining_val = {
+                        let mut fns_remaining_ref = fns_remaining.write().await;
+                        *fns_remaining_ref -= 1;
+                        *fns_remaining_ref
+                    };
+                    if fns_remaining_val == 0 {
+                        fn_done_tx.write().await.take();
+                    }
+                })
+                .await;
+
+            drop(result_tx);
+        };
+
+        futures::join!(queuer, scheduler);
+
+        let results = stream::poll_fn(move |ctx| result_rx.poll_recv(ctx))
+            .collect::<Vec<E>>()
+            .await;
+
+        if results.is_empty() {
+            Ok(())
+        } else {
+            Err(results)
+        }
+    }
+
+    /// Runs the provided logic over the functions concurrently in topological
+    /// order, stopping when an error is encountered.
+    ///
+    /// This gracefully waits until all produced tasks have returned. The return
+    /// value is a `Vec<E>` as it is possible for multiple tasks to return
+    /// errors.
+    ///
+    /// The first argument is an optional `limit` on the number of concurrent
+    /// futures. If this limit is not `None`, no more than `limit` futures will
+    /// be run concurrently. The `limit` argument is of type
+    /// `Into<Option<usize>>`, and so can be provided as either `None`,
+    /// `Some(10)`, or just `10`.
+    ///
+    /// **Note:** a limit of zero is interpreted as no limit at all, and will
+    /// have the same result as passing in `None`.
+    #[cfg(feature = "async")]
+    pub async fn try_for_each_concurrent_mut<E, FnTryForEach, Fut>(
+        &mut self,
+        limit: impl Into<Option<usize>>,
+        fn_try_for_each: FnTryForEach,
+    ) -> Result<(), Vec<E>>
+    where
+        E: Debug,
+        FnTryForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = Result<(), E>>,
+    {
+        self.try_for_each_concurrent_mut_internal(limit, fn_try_for_each, IterDirection::Forward)
+            .await
+    }
+
+    /// Runs the provided logic over the functions concurrently in reverse
+    /// topological order, stopping when an error is encountered.
+    ///
+    /// This gracefully waits until all produced tasks have returned. The return
+    /// value is a `Vec<E>` as it is possible for multiple tasks to return
+    /// errors.
+    ///
+    /// The first argument is an optional `limit` on the number of concurrent
+    /// futures. If this limit is not `None`, no more than `limit` futures will
+    /// be run concurrently. The `limit` argument is of type
+    /// `Into<Option<usize>>`, and so can be provided as either `None`,
+    /// `Some(10)`, or just `10`.
+    ///
+    /// **Note:** a limit of zero is interpreted as no limit at all, and will
+    /// have the same result as passing in `None`.
+    #[cfg(feature = "async")]
+    pub async fn try_for_each_concurrent_mut_rev<E, FnTryForEach, Fut>(
+        &mut self,
+        limit: impl Into<Option<usize>>,
+        fn_try_for_each: FnTryForEach,
+    ) -> Result<(), Vec<E>>
+    where
+        E: Debug,
+        FnTryForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = Result<(), E>>,
+    {
+        self.try_for_each_concurrent_mut_internal(limit, fn_try_for_each, IterDirection::Reverse)
+            .await
+    }
+
+    // https://users.rust-lang.org/t/lifetime-may-not-live-long-enough-for-an-async-closure/62489
+    #[cfg(feature = "async")]
+    async fn try_for_each_concurrent_mut_internal<E, FnTryForEach, Fut>(
         &mut self,
         limit: impl Into<Option<usize>>,
         fn_try_for_each: FnTryForEach,
@@ -520,7 +796,8 @@ impl<F> FnGraph<F> {
     ) -> Result<(), Vec<E>>
     where
         E: Debug,
-        FnTryForEach: Fn(&mut F) -> BoxFuture<'_, Result<(), E>>,
+        FnTryForEach: Fn(&mut F) -> Fut,
+        Fut: Future<Output = Result<(), E>>,
     {
         let (graph_structure, predecessor_counts) = match iter_direction {
             IterDirection::Forward => (&self.graph_structure, self.edge_counts.incoming().to_vec()),
@@ -983,9 +1260,9 @@ mod tests {
                 test_timeout(
                     Duration::from_millis(200),
                     Duration::from_millis(255),
-                    fn_graph
-                        .stream()
-                        .for_each_concurrent(None, |f| async move { f.call(resources).await }),
+                    fn_graph.stream().for_each_concurrent(None, |f| async move {
+                        let _ = f.call(resources).await;
+                    }),
                 )
                 .await;
 
@@ -1017,7 +1294,9 @@ mod tests {
                     Duration::from_millis(255),
                     fn_graph
                         .stream_rev()
-                        .for_each_concurrent(None, |f| async move { f.call(resources).await }),
+                        .for_each_concurrent(None, |f| async move {
+                            let _ = f.call(resources).await;
+                        }),
                 )
                 .await;
 
@@ -1097,13 +1376,12 @@ mod tests {
         }
 
         #[test]
-        fn for_each_concurrent_runs_fns_concurrently_mut() -> Result<(), Box<dyn std::error::Error>>
-        {
+        fn for_each_concurrent_runs_fns_concurrently() -> Result<(), Box<dyn std::error::Error>> {
             let rt = runtime::Builder::new_current_thread()
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1113,15 +1391,15 @@ mod tests {
                     Duration::from_millis(200),
                     Duration::from_millis(255),
                     fn_graph.for_each_concurrent(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             let _ = fut.await;
                         }
-                        .boxed()
                     }),
                 )
                 .await;
 
+                seq_rx.close();
                 let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
                     .collect::<Vec<&'static str>>()
                     .await;
@@ -1132,13 +1410,13 @@ mod tests {
         }
 
         #[test]
-        fn for_each_concurrent_rev_runs_fns_concurrently_mut()
-        -> Result<(), Box<dyn std::error::Error>> {
+        fn for_each_concurrent_rev_runs_fns_concurrently() -> Result<(), Box<dyn std::error::Error>>
+        {
             let rt = runtime::Builder::new_current_thread()
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1148,15 +1426,15 @@ mod tests {
                     Duration::from_millis(200),
                     Duration::from_millis(255),
                     fn_graph.for_each_concurrent_rev(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             let _ = fut.await;
                         }
-                        .boxed()
                     }),
                 )
                 .await;
 
+                seq_rx.close();
                 let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
                     .collect::<Vec<&'static str>>()
                     .await;
@@ -1167,13 +1445,13 @@ mod tests {
         }
 
         #[test]
-        fn try_for_each_concurrent_runs_fns_concurrently_mut()
-        -> Result<(), Box<dyn std::error::Error>> {
+        fn try_for_each_concurrent_runs_fns_concurrently() -> Result<(), Box<dyn std::error::Error>>
+        {
             let rt = runtime::Builder::new_current_thread()
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1184,17 +1462,17 @@ mod tests {
                     Duration::from_millis(200),
                     Duration::from_millis(255),
                     fn_graph.try_for_each_concurrent(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             let _ = fut.await;
                             Result::<_, TestError>::Ok(())
                         }
-                        .boxed()
                     }),
                 )
                 .await
                 .unwrap();
 
+                seq_rx.close();
                 let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
                     .collect::<Vec<&'static str>>()
                     .await;
@@ -1212,7 +1490,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1223,14 +1501,13 @@ mod tests {
                     Duration::from_millis(50),
                     Duration::from_millis(70),
                     fn_graph.try_for_each_concurrent(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             match fut.await {
                                 "a" => Err(TestError("a")),
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
                     }),
                 )
                 .await;
@@ -1251,7 +1528,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1262,14 +1539,13 @@ mod tests {
                     Duration::from_millis(100),
                     Duration::from_millis(120),
                     fn_graph.try_for_each_concurrent(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             match fut.await {
                                 "c" => Err(TestError("c")),
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
                     }),
                 )
                 .await;
@@ -1292,7 +1568,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1303,7 +1579,7 @@ mod tests {
                     Duration::from_millis(100),
                     Duration::from_millis(120),
                     fn_graph.try_for_each_concurrent(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             match fut.await {
                                 "b" => Err(TestError("b")),
@@ -1311,7 +1587,6 @@ mod tests {
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
                     }),
                 )
                 .await;
@@ -1333,13 +1608,13 @@ mod tests {
         }
 
         #[test]
-        fn try_for_each_concurrent_rev_runs_fns_concurrently_mut()
+        fn try_for_each_concurrent_rev_runs_fns_concurrently()
         -> Result<(), Box<dyn std::error::Error>> {
             let rt = runtime::Builder::new_current_thread()
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1350,17 +1625,17 @@ mod tests {
                     Duration::from_millis(200),
                     Duration::from_millis(255),
                     fn_graph.try_for_each_concurrent_rev(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             let _ = fut.await;
                             Result::<_, TestError>::Ok(())
                         }
-                        .boxed()
                     }),
                 )
                 .await
                 .unwrap();
 
+                seq_rx.close();
                 let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
                     .collect::<Vec<&'static str>>()
                     .await;
@@ -1378,7 +1653,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1389,14 +1664,13 @@ mod tests {
                     Duration::from_millis(50),
                     Duration::from_millis(70),
                     fn_graph.try_for_each_concurrent_rev(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             match fut.await {
                                 "e" => Err(TestError("e")),
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
                     }),
                 )
                 .await;
@@ -1416,7 +1690,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1427,14 +1701,13 @@ mod tests {
                     Duration::from_millis(150),
                     Duration::from_millis(190),
                     fn_graph.try_for_each_concurrent_rev(None, |f| {
-                        let fut = f.call_mut(resources);
+                        let fut = f.call(resources);
                         async move {
                             match fut.await {
                                 "b" => Err(TestError("b")),
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
                     }),
                 )
                 .await;
@@ -1457,7 +1730,7 @@ mod tests {
                 .enable_time()
                 .build()?;
             rt.block_on(async {
-                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+                let (fn_graph, mut seq_rx) = complex_graph_unit()?;
 
                 let mut resources = Resources::new();
                 resources.insert(0u8);
@@ -1468,6 +1741,234 @@ mod tests {
                     Duration::from_millis(150),
                     Duration::from_millis(190),
                     fn_graph.try_for_each_concurrent_rev(None, |f| {
+                        let fut = f.call(resources);
+                        async move {
+                            match fut.await {
+                                "b" => Err(TestError("b")),
+                                "c" => Err(TestError("c")),
+                                _ => Ok(()),
+                            }
+                        }
+                    }),
+                )
+                .await;
+
+                assert_eq!(
+                    [TestError("b"), TestError("c")],
+                    result.unwrap_err().as_slice()
+                );
+                assert_eq!("e", seq_rx.try_recv().unwrap());
+                assert_eq!("d", seq_rx.try_recv().unwrap());
+                assert_eq!("b", seq_rx.try_recv().unwrap());
+                assert_eq!("c", seq_rx.try_recv().unwrap());
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn for_each_concurrent_mut_runs_fns_concurrently() -> Result<(), Box<dyn std::error::Error>>
+        {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+                test_timeout(
+                    Duration::from_millis(200),
+                    Duration::from_millis(255),
+                    fn_graph.for_each_concurrent_mut(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            let _ = fut.await;
+                        }
+                    }),
+                )
+                .await;
+
+                let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
+                    .collect::<Vec<&'static str>>()
+                    .await;
+
+                assert_eq!(["f", "a", "c", "b", "d", "e"], fn_iter_order.as_slice());
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn for_each_concurrent_mut_rev_runs_fns_concurrently()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+                test_timeout(
+                    Duration::from_millis(200),
+                    Duration::from_millis(255),
+                    fn_graph.for_each_concurrent_mut_rev(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            let _ = fut.await;
+                        }
+                    }),
+                )
+                .await;
+
+                let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
+                    .collect::<Vec<&'static str>>()
+                    .await;
+
+                assert_eq!(["e", "d", "b", "c", "f", "a"], fn_iter_order.as_slice());
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_runs_fns_concurrently_mut()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                test_timeout(
+                    Duration::from_millis(200),
+                    Duration::from_millis(255),
+                    fn_graph.try_for_each_concurrent_mut(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            let _ = fut.await;
+                            Result::<_, TestError>::Ok(())
+                        }
+                    }),
+                )
+                .await
+                .unwrap();
+
+                let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
+                    .collect::<Vec<&'static str>>()
+                    .await;
+
+                assert_eq!(["f", "a", "c", "b", "d", "e"], fn_iter_order.as_slice());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_gracefully_ends_when_one_function_returns_failure()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(50),
+                    Duration::from_millis(70),
+                    fn_graph.try_for_each_concurrent_mut(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            match fut.await {
+                                "a" => Err(TestError("a")),
+                                _ => Ok(()),
+                            }
+                        }
+                    }),
+                )
+                .await;
+
+                assert_eq!([TestError("a")], result.unwrap_err().as_slice());
+                assert_eq!("f", seq_rx.try_recv().unwrap());
+                assert_eq!("a", seq_rx.try_recv().unwrap()); // "a" is sent before we err
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_gracefully_ends_when_one_function_returns_failure_variation()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(100),
+                    Duration::from_millis(120),
+                    fn_graph.try_for_each_concurrent_mut(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            match fut.await {
+                                "c" => Err(TestError("c")),
+                                _ => Ok(()),
+                            }
+                        }
+                    }),
+                )
+                .await;
+
+                assert_eq!([TestError("c")], result.unwrap_err().as_slice());
+                assert_eq!("f", seq_rx.try_recv().unwrap());
+                assert_eq!("a", seq_rx.try_recv().unwrap());
+                assert_eq!("c", seq_rx.try_recv().unwrap());
+                assert_eq!("b", seq_rx.try_recv().unwrap());
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_gracefully_ends_when_multiple_functions_return_failure()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(100),
+                    Duration::from_millis(120),
+                    fn_graph.try_for_each_concurrent_mut(None, |f| {
                         let fut = f.call_mut(resources);
                         async move {
                             match fut.await {
@@ -1476,7 +1977,167 @@ mod tests {
                                 _ => Ok(()),
                             }
                         }
-                        .boxed()
+                    }),
+                )
+                .await;
+
+                // Both "c" and "b" being present proves we have waited for in-progress tasks to
+                // complete.
+                assert_eq!(
+                    [TestError("c"), TestError("b")],
+                    result.unwrap_err().as_slice()
+                );
+                assert_eq!("f", seq_rx.try_recv().unwrap());
+                assert_eq!("a", seq_rx.try_recv().unwrap());
+                assert_eq!("c", seq_rx.try_recv().unwrap());
+                assert_eq!("b", seq_rx.try_recv().unwrap());
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_rev_runs_fns_concurrently_mut()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                test_timeout(
+                    Duration::from_millis(200),
+                    Duration::from_millis(255),
+                    fn_graph.try_for_each_concurrent_mut_rev(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            let _ = fut.await;
+                            Result::<_, TestError>::Ok(())
+                        }
+                    }),
+                )
+                .await
+                .unwrap();
+
+                let fn_iter_order = stream::poll_fn(|context| seq_rx.poll_recv(context))
+                    .collect::<Vec<&'static str>>()
+                    .await;
+
+                assert_eq!(["e", "d", "b", "c", "f", "a"], fn_iter_order.as_slice());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_rev_gracefully_ends_when_one_function_returns_failure()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(50),
+                    Duration::from_millis(70),
+                    fn_graph.try_for_each_concurrent_mut_rev(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            match fut.await {
+                                "e" => Err(TestError("e")),
+                                _ => Ok(()),
+                            }
+                        }
+                    }),
+                )
+                .await;
+
+                assert_eq!([TestError("e")], result.unwrap_err().as_slice());
+                assert_eq!("e", seq_rx.try_recv().unwrap()); // "a" is sent before we err
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_rev_gracefully_ends_when_one_function_returns_failure_variation()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(150),
+                    Duration::from_millis(190),
+                    fn_graph.try_for_each_concurrent_mut_rev(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            match fut.await {
+                                "b" => Err(TestError("b")),
+                                _ => Ok(()),
+                            }
+                        }
+                    }),
+                )
+                .await;
+
+                assert_eq!([TestError("b")], result.unwrap_err().as_slice());
+                assert_eq!("e", seq_rx.try_recv().unwrap());
+                assert_eq!("d", seq_rx.try_recv().unwrap());
+                assert_eq!("b", seq_rx.try_recv().unwrap());
+                assert_eq!("c", seq_rx.try_recv().unwrap());
+                assert_eq!(TryRecvError::Empty, seq_rx.try_recv().unwrap_err());
+
+                Ok(())
+            })
+        }
+
+        #[test]
+        fn try_for_each_concurrent_mut_rev_gracefully_ends_when_multiple_functions_return_failure()
+        -> Result<(), Box<dyn std::error::Error>> {
+            let rt = runtime::Builder::new_current_thread()
+                .enable_time()
+                .build()?;
+            rt.block_on(async {
+                let (mut fn_graph, mut seq_rx) = complex_graph_unit_mut()?;
+
+                let mut resources = Resources::new();
+                resources.insert(0u8);
+                resources.insert(0u16);
+                let resources = &resources;
+
+                let result = test_timeout(
+                    Duration::from_millis(150),
+                    Duration::from_millis(190),
+                    fn_graph.try_for_each_concurrent_mut_rev(None, |f| {
+                        let fut = f.call_mut(resources);
+                        async move {
+                            match fut.await {
+                                "b" => Err(TestError("b")),
+                                "c" => Err(TestError("c")),
+                                _ => Ok(()),
+                            }
+                        }
                     }),
                 )
                 .await;
@@ -1509,12 +2170,14 @@ mod tests {
             // concurrent
             assert!(
                 duration_elapsed < upper,
-                "duration was {} ms",
+                "expected duration to be less than {} ms, duration was {} ms",
+                upper.as_millis(),
                 duration_elapsed.as_millis()
             );
             assert!(
                 duration_elapsed > lower,
-                "duration was {} ms",
+                "expected duration to be more than {} ms, duration was {} ms",
+                lower.as_millis(),
                 duration_elapsed.as_millis()
             );
 
@@ -1523,7 +2186,7 @@ mod tests {
 
         fn complex_graph_unit() -> Result<
             (
-                FnGraph<Box<dyn FnRes<Ret = BoxFuture<'static, ()>>>>,
+                FnGraph<Box<dyn FnRes<Ret = BoxFuture<'static, &'static str>>>>,
                 Receiver<&'static str>,
             ),
             WouldCycle<Edge>,
@@ -1551,62 +2214,68 @@ mod tests {
             let mut fn_graph_builder = FnGraphBuilder::new();
             let [fn_id_a, fn_id_b, fn_id_c, fn_id_d, fn_id_e, fn_id_f] =
                 fn_graph_builder.add_fns([
-                    (move |_: &u8| -> BoxFuture<'_, ()> {
+                    (move |_: &u8| -> BoxFuture<'_, &'static str> {
                         seq_tx_a
                             .try_send("a")
                             .expect("Failed to send sequence `a`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "a"
                         }
                         .boxed()
                     })
                     .into_fn_res(),
-                    (move |_: &mut u16| -> BoxFuture<'_, ()> {
+                    (move |_: &mut u16| -> BoxFuture<'_, &'static str> {
                         seq_tx_b
                             .try_send("b")
                             .expect("Failed to send sequence `b`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "b"
                         }
                         .boxed()
                     })
                     .into_fn_res(),
-                    (move || -> BoxFuture<'_, ()> {
+                    (move || -> BoxFuture<'_, &'static str> {
                         seq_tx_c
                             .try_send("c")
                             .expect("Failed to send sequence `c`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "c"
                         }
                         .boxed()
                     })
                     .into_fn_res(),
-                    (move |_: &u8, _: &mut u16| -> BoxFuture<'_, ()> {
+                    (move |_: &u8, _: &mut u16| -> BoxFuture<'_, &'static str> {
                         seq_tx_d
                             .try_send("d")
                             .expect("Failed to send sequence `d`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "d"
                         }
                         .boxed()
                     })
                     .into_fn_res(),
-                    (move || -> BoxFuture<'_, ()> {
+                    (move || -> BoxFuture<'_, &'static str> {
                         seq_tx_e
                             .try_send("e")
                             .expect("Failed to send sequence `e`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "e"
                         }
                         .boxed()
                     })
                     .into_fn_res(),
-                    (move |_: &mut u16| -> BoxFuture<'_, ()> {
+                    (move |_: &mut u16| -> BoxFuture<'_, &'static str> {
                         seq_tx_f
                             .try_send("f")
                             .expect("Failed to send sequence `f`.");
                         async {
                             time::sleep(sleep_duration!()).await;
+                            "f"
                         }
                         .boxed()
                     })